API client configuration
