+++ src/Lewee.Blazor/Lewee.Blazor.csproj
@@ -9,6 +9,7 @@
     <PackageReference Include="Flurl" />
     <PackageReference Include="Fluxor.Blazor.Web.ReduxDevTools" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" />
+    <PackageReference Include="Refit.HttpClientFactory" />
   </ItemGroup>
 
   <ItemGroup>
